+++ src/lib/react-query.ts
@@ -1,5 +1,10 @@
-import { QueryClient, DefaultOptions, UseQueryOptions, UseMutationOptions } from '@tanstack/react-query';
 import { AxiosError } from 'axios';
+import { 
+  QueryClient,
+  DefaultOptions,
+  UseQueryOptions,
+  UseMutationOptions
+} from '@tanstack/react-query';
 
 const queryConfig: DefaultOptions= {
   queries: {
