+++ src/routes/index.tsx
@@ -1,22 +1,25 @@
 import { useRoutes }      from 'react-router-dom';
 
-import { Landing } from '@/features/misc';
-
 import { authorizedRoutes } from './authorizedRoutes';
 import { publicRoutes }     from './publicRoutes';
-import { AuthenticatedResponseType, IS_LOGGED_IN_KEY, useAuthenticatedUserQuery } from '@/features/auth';
-import { ModeSelection } from '@/features/modeSelect';
+import { Loading }          from '@/components/Animation';
+import { 
+  AuthenticatedResponseType,
+  IS_LOGGED_IN_KEY,
+  useAuthenticatedUserQuery
+} from '@/features/auth';
+
 
 
 export const AppRoutes = () => {
   const { data: isAuthenticated, isLoading }= useAuthenticatedUserQuery({
     config: {
-      select: ( authUser:AuthenticatedResponseType ) => authUser[ IS_LOGGED_IN_KEY ]
+      select: ( authUser: AuthenticatedResponseType ) => authUser[ IS_LOGGED_IN_KEY ]
     }
   });
 
   if( isLoading ) {
-    return <>loading</>
+    return <Loading />
   };
 
   const routes = isAuthenticated ? authorizedRoutes : publicRoutes;
