+++ src/routes/publicRoutes.tsx
@@ -1,9 +1,9 @@
-import { lazyImport } from "@/utils/lazyImport";
-import { Navigate } from "react-router";
+import { Navigate } from 'react-router';
 
+import { lazyImport } from '@/utils/lazyImport';
 const { AuthRoutes }= lazyImport(() => import('@/features/auth'), 'AuthRoutes');
 
 export const publicRoutes = [
   { path: 'auth/*', element: <AuthRoutes /> },
-  { path: '*', element: <Navigate to='auth/login' /> },
+  { path: '*',      element: <Navigate to='auth/login' /> },
 ];
